# openapi.yaml
openapi: 3.0.0
paths:
...
  /employees:
    post:
      description: Create a new employee record in the database with name, email, and role
      operationId: EmployeesController_create
      parameters: []
      requestBody:
        required: true
        description: Employee data to create
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEmployeeDto"
      responses:
        "201":
          description: Employee created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmployeeResponseDto"
        "400":
          description: Invalid input data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseDto"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseDto"
        "409":
          description: Conflict (e.g., duplicate entry)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseDto"
        "503":
          description: Service unavailable (e.g., database connection error)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseDto"
      security:
        - session-id: []
      summary: Create a new employee
      tags: &a3
        - employees
    get:
      description: Retrieve a list of all employees from the database. Supports filtering by role, searching by name/email, and sorting by createdAt or name.
      operationId: EmployeesController_findAll
      parameters:
        - name: role
          required: false
          in: query
          description: Filter employees by role
          schema:
            enum:
              - INTERN
              - ENGINEER
              - ADMIN
            type: string
        - name: page
          required: false
          in: query
          description: "Page number (default: 1)"
          schema:
            example: 1
            type: number
        - name: pageSize
          required: false
          in: query
          description: "Number of items per page (default: 10)"
          schema:
            example: 10
            type: number
        - name: searchName
          required: false
          in: query
          description: Search employees by name (partial match)
          schema:
            example: John
            type: string
        - name: searchEmail
          required: false
          in: query
          description: Search employees by email (partial match)
          schema:
            example: john@example.com
            type: string
        - name: sortBy
          required: false
          in: query
          description: "Column to sort by (default: id)"
          schema:
            enum:
              - createdAt
              - name
            type: string
        - name: sortOrder
          required: false
          in: query
          description: "Sort order (default: ASC)"
          schema:
            enum:
              - ASC
              - DESC
            type: string
        - name: X-Has-Previous-Page
          in: header
          description: Whether there is a previous page (true/false)
          schema:
            type: string
        - name: X-Has-Next-Page
          in: header
          description: Whether there is a next page (true/false)
          schema:
            type: string
        - name: X-Total-Pages
          in: header
          description: Total pages available for the current query
          schema:
            type: string
        - name: X-Page-Size
          in: header
          description: Number of items per page
          schema:
            type: string
        - name: X-Page
          in: header
          description: Current page number
          schema:
            type: string
        - name: X-Total-Count
          in: header
          description: Total number of matching records
          schema:
            type: string
      responses:
        "200":
          description: List of employees for the current page. Pagination metadata is returned via response headers.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/EmployeeResponseDto"
        "400":
          description: Invalid query parameters
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseDto"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseDto"
        "403":
          description: Forbidden (insufficient permissions)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseDto"
        "503":
          description: Service unavailable (e.g., database connection error)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseDto"
      security:
        - session-id: []
      summary: Get all employees
      tags: *a3
  /employees/{id}:
    get:
      description: Retrieve a specific employee by their unique identifier from the database
      operationId: EmployeesController_findOne
      parameters:
        - name: id
          required: true
          in: path
          description: Employee ID
          schema:
            example: "1"
            type: string
      responses:
        "200":
          description: Employee retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmployeeResponseDto"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseDto"
        "403":
          description: Forbidden (insufficient permissions)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseDto"
        "404":
          description: Employee not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseDto"
        "503":
          description: Service unavailable (e.g., database connection error)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseDto"
      security:
        - session-id: []
      summary: Get employee by ID
      tags: *a3
    patch:
      description: Update an existing employee record. Only provided fields will be updated.
      operationId: EmployeesController_update
      parameters:
        - name: id
          required: true
          in: path
          description: Employee ID
          schema:
            example: "1"
            type: string
      requestBody:
        required: true
        description: Employee data to update (partial)
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateEmployeeDto"
      responses:
        "200":
          description: Employee updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmployeeResponseDto"
        "400":
          description: Invalid input data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseDto"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseDto"
        "403":
          description: Forbidden (insufficient permissions)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseDto"
        "404":
          description: Employee not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseDto"
        "409":
          description: Conflict (e.g., duplicate entry)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseDto"
        "503":
          description: Service unavailable (e.g., database connection error)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseDto"
      security:
        - session-id: []
      summary: Update employee
      tags: *a3
    delete:
      description: Delete an employee record from the database by their unique identifier
      operationId: EmployeesController_remove
      parameters:
        - name: id
          required: true
          in: path
          description: Employee ID
          schema:
            example: "1"
            type: string
      responses:
        "200":
          description: Employee deleted successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmployeeResponseDto"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseDto"
        "403":
          description: Forbidden (insufficient permissions)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseDto"
        "404":
          description: Employee not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseDto"
        "503":
          description: Service unavailable (e.g., database connection error)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseDto"
      summary: Delete employee
      tags: *a3
  /employees/{id}/photo:
    post:
      description: Upload a profile photo for an employee. Accepts JPEG, PNG, or WebP images up to 5MB.
      operationId: EmployeesController_uploadPhoto
      parameters:
        - name: id
          required: true
          in: path
          description: Employee ID
          schema:
            example: "1"
            type: string
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                  description: Image file (JPEG, PNG, or WebP, max 5MB)
      responses:
        "200":
          description: Photo uploaded successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmployeeResponseDto"
        "400":
          description: Invalid file type or size
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseDto"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseDto"
        "403":
          description: Forbidden (insufficient permissions)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseDto"
        "404":
          description: Employee not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseDto"
        "503":
          description: Service unavailable (e.g., storage/database error)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseDto"
      security:
        - session-id: []
      summary: Upload employee photo
      tags: *a3
    delete:
      description: Remove the profile photo for an employee
      operationId: EmployeesController_deletePhoto
      parameters:
        - name: id
          required: true
          in: path
          description: Employee ID
          schema:
            example: "1"
            type: string
      responses:
        "200":
          description: Photo deleted successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmployeeResponseDto"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseDto"
        "403":
          description: Forbidden (insufficient permissions)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseDto"
        "404":
          description: Employee not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseDto"
        "503":
          description: Service unavailable (e.g., storage/database error)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseDto"
      security:
        - session-id: []
      summary: Delete employee photo
      tags: *a3
...
info:
  title: Nest API Demo
  description: A comprehensive REST API built with NestJS and MySQL. This API provides endpoints for managing users, employees, and departments with role-based access control.
  version: "1.0"
  contact: {}
...
servers: []
components:
  securitySchemes:
    cookie:
      type: apiKey
      in: cookie
      name: session-id
  schemas:
...
    ErrorResponseDto:
      type: object
      properties:
        statusCode:
          type: number
          description: HTTP status code of the error response
          example: 401
        timestamp:
          type: string
          description: ISO 8601 timestamp of when the error occurred
          example: 2024-01-15T10:30:00.000Z
          format: date-time
        path:
          type: string
          description: Request path that produced the error
          example: /api/employees/123
        response:
          description: Error response payload (string message or an object returned by Nest exceptions)
          oneOf:
            - type: string
              example: Unauthorized
            - type: object
              additionalProperties: true
              example:
                message: Unauthorized
                statusCode: 401
      required:
        - statusCode
        - timestamp
        - path
        - response
...
    CreateEmployeeDto:
      type: object
      properties:
        name:
          type: string
          description: The full name of the employee
          example: Jane Smith
          minLength: 1
        email:
          type: string
          description: The email address of the employee
          example: jane.smith@example.com
          format: email
        role:
          type: string
          description: The role of the employee
          enum: &a7
            - INTERN
            - ENGINEER
            - ADMIN
          example: ENGINEER
        photoUrl:
          type: object
          description: The URL of the employee profile photo
          example: https://namespace.compat.objectstorage.region.oraclecloud.com/bucket/employees/photo.jpg
          nullable: true
      required:
        - name
        - email
        - role
    EmployeeResponseDto:
      type: object
      properties:
        id:
          type: number
          description: The unique identifier of the employee
          example: 1
        name:
          type: string
          description: The full name of the employee
          example: Jane Smith
        email:
          type: string
          description: The email address of the employee
          example: jane.smith@example.com
          format: email
        role:
          type: string
          description: The role of the employee
          enum:
            - INTERN
            - ENGINEER
            - ADMIN
          example: ENGINEER
        photoUrl:
          type: object
          description: The URL of the employee profile photo
          example: https://namespace.compat.objectstorage.region.oraclecloud.com/bucket/employees/photo.jpg
          nullable: true
        createdAt:
          type: string
          description: The date and time when the employee was created
          example: 2024-01-15T10:30:00.000Z
          format: date-time
        updatedAt:
          type: string
          description: The date and time when the employee was last updated
          example: 2024-01-15T10:30:00.000Z
          format: date-time
      required:
        - id
        - name
        - email
        - role
        - createdAt
        - updatedAt
    UpdateEmployeeDto:
      type: object
      properties:
        name:
          type: string
          description: The full name of the employee
          example: Jane Smith
          minLength: 1
        email:
          type: string
          description: The email address of the employee
          example: jane.smith@example.com
          format: email
        role:
          type: string
          description: The role of the employee
          enum: *a7
          example: ENGINEER
        photoUrl:
          type: object
          description: The URL of the employee profile photo
          example: https://namespace.compat.objectstorage.region.oraclecloud.com/bucket/employees/photo.jpg
          nullable: true
...